requirements.txt
ninja=1.9.0
python=3.7.4
pytorch=1.2.0
scikit-image=0.15.0
scikit-learn=0.21.3
scipy=1.4.1
seaborn=0.9.0
tensorboardx=2.0
tensorflow=2.1.0
torchvision=0.4.0
tqdm=4.36.1